SimplexGridFactory — Module

SimplexGridFactory

SimplexGridFactory

Provides SimplexGridBuilder, a convenience interface to rhe triangle (via Triangulate.jl) and TetGen (via TetGen.jl) mesh generators. Code using it needs to pass the Triangulate resp. the TetGen module as parameter to the constructor of SimplexGridBuilder and thus needs to adhere to their respective license conditions. This package itself is MIT licensed.

Recent changes

Please look up the list of recent changes

API Index

Builder API

Grid generator options

Test plots

Some primitives

Changes

v2.0 March 21, 2024

v1.0 Nov 15, 2023

Mesh generator interface

Pluto Notebook(s)

cylinder.jl

gridgenvis.jl

2D Triangulations

Domain triangulation

Triangulation with size control and boundary markers

Triangulation with subregions

Direct specification of input arrays

Local refinement

Domain with holes

Glueing in another grid

3D Tetrahedralizations

Domain triangulation

Cube based on primitves

Glue-in of existing grid

Load STL model