Mesh generator interface

ExtendableGrids.simplexgrid — Method

simplexgrid(
    ::Type{SimplexGridFactory.TriangulateType},
    Triangulate,
    input;
    kwargs...
) -> ExtendableGrid

Create Grid from Triangle input data.

See default_options for available kwargs.

SimplexGridFactory.triangulateio — Method

triangulateio(
    Triangulate;
    points,
    bfaces,
    bfaceregions,
    regionpoints,
    regionnumbers,
    regionvolumes
) -> Any

Create a TriangulateIO structure from a number of input arrays. The 2D input arrays are transposed if necessary and converted to the proper data types for Triangulate.

This conversion is not performed if the data types are those indicated in the defaults and the leading dimension of 2D arrays corresponds to the space dimension.

SimplexGridFactory.triangulateio — Method

triangulateio(this::SimplexGridBuilder) -> Any

Create triangle input from the current state of the builder.

ExtendableGrids.simplexgrid — Method

simplexgrid(
    ::Type{SimplexGridFactory.TetGenType},
    TetGen,
    input;
    kwargs...
) -> ExtendableGrid

Create Grid from TetGen data.

See default_options for available kwargs.

SimplexGridFactory.tetgenio — Method

tetgenio(
    TetGen;
    points,
    bfaces,
    bfaceregions,
    regionpoints,
    regionnumbers,
    regionvolumes
) -> Any

Create a RawTetGenIO structure from a number of input arrays. The 2D input arrays are transposed if necessary and converted to the proper data types for TetGen.

This conversion is not performed if the data types are those indicated in the defaults and the leading dimension of 2D arrays corresponds to the space dimension.

SimplexGridFactory.tetgenio — Method

tetgenio(this::SimplexGridBuilder) -> Any

Create tetgen input from the current state of the builder.

ExtendableGrids.simplexgrid — Method

      simplexgrid(Generator;
                     points=Array{Cdouble,2}(undef,0,0),
                     bfaces=Array{Cint,2}(undef,0,0),
                     bfaceregions=Array{Cint,1}(undef,0),
                     regionpoints=Array{Cdouble,2}(undef,0,0),
                     regionnumbers=Array{Cint,1}(undef,0),
                     regionvolumes=Array{Cdouble,1}(undef,0);
                     kwargs...
                  )

Create Grid from a number of input arrays. The 2D input arrays are transposed if necessary and converted to the proper data types for Triangulate or TetGen

This conversion is not performed if the data types are those indicated in the defaults and the leading dimension of 2D arrays corresponds to the space dimension.

See default_options for available kwargs.

SimplexGridFactory.istriangulate — Function

istriangulate(Generator)

Check if generator is Triangulate.

SimplexGridFactory.istetgen — Function

istetgen(Generator)

Check if generator is TetGen.